Merging

+++ Player.cpp
@@ -112,6 +112,6 @@ void Player::check(){
     if(zpos < -9.74) zpos = -9.74;
 }
 
-int Player::getX(){return xpos;}
-int Player::getZ(){return zpos;}
-int Player::getY(){return yrot;}
+GLfloat Player::getX(){return xpos;}
+GLfloat Player::getZ(){return zpos;}
+GLfloat Player::getY(){return yrot;}
